Use the --production flag on 'npm audit'
npm audit was broken as the 'default' working-directory
and the run steps working-directory were in conflict. The
existing 'npm audit' didn't actually run but was hidden by the
'continue-on-error' property. This change fixes the working directory
and replaces the use of modifying the package.json file with the
npm audit '--production' flag to prevent scanning the devDependencies.

Signed-off-by: Brett Logan <lindluni@github.com>